scripts/requirements.txt
yfinance
pyyaml
pandas
numpy
